Split boxmesh and libmesh remote dof tests
