+++ src/discretization/test/CMakeLists.txt
@@ -17,12 +17,13 @@ ENDIF()
 ADD_AMP_TEST_1_2_4( test_Discretization )
 ADD_AMP_TEST_1_2_4( test_Neighbors )
 
-ADD_AMP_TEST_1_2_4( test_RemoteDOFs )
+ADD_AMP_TEST_1_2_4( test_RemoteDOFs_BoxMesh )
 COPY_TEST_FILE( input_testRemoteDOFs-boxmesh-1 )
 COPY_TEST_FILE( input_testRemoteDOFs-boxmesh-2 )
-COPY_TEST_FILE( input_testRemoteDOFs-libmesh-1 )
-COPY_TEST_FILE( input_testRemoteDOFs-libmesh-2 )
 
 IF ( USE_LIBMESH )
     ADD_AMP_TEST_1_2_4( test_libmeshElement )
+    ADD_AMP_TEST_1_2_4( test_RemoteDOFs_LibMesh )
+    COPY_TEST_FILE( input_testRemoteDOFs-libmesh-1 )
+    COPY_TEST_FILE( input_testRemoteDOFs-libmesh-2 )
 ENDIF()
